int64

+++ linbox/tests/test-solve.C
@@ -770,7 +770,7 @@ int main (int argc, char **argv)
 		END_OF_ARGUMENTS
 	};
 
-	typedef Givaro::Modular<uint32_t> Field;
+	typedef Givaro::Modular<uint64_t> Field;
 
 	parseArguments (argc, argv, args);
 	Field F (q);
